Reorganize frontend sources under public libreria
