+++ public/index.html
@@ -12,6 +12,6 @@
         <div id="messages"></div>
         <main id="main"></main>
     </div>
-    <script type="module" src="/app/main/app.js"></script>
+    <script type="module" src="/libreria/js/main.mjs"></script>
 </body>
 </html>
