+++ public/libreria/js/common/libreria-session.js
