+++ public/libreria/js/common/presenter.js
