+++ public/libreria/js/components/admin/libros.js
