+++ public/libreria/js/components/admin/ver-libro.js
