+++ public/libreria/js/components/cliente/ver-libro.js
