+++ public/libreria/js/components/invitado/registro.js
