+++ public/libreria/js/components/invitado/ver-libro.js
