+++ public/libreria/js/model/auth-store.js
