+++ public/libreria/js/model/libro.js
